Specify /usr/racket install directory.
Although that's the default for release builds, the nightly builds
default to e.g. /usr/racket-5.90.0.9. Although that default makes sense
where people are doing side-by-side installs, we're doing an install on
a throw-away Travis VM. More basically, we need the build scripts to
find stuff consistently in /usr/racket regardless of whether this is a
release or HEAD.