Try getting install-racket.sh using git clone instead of curl
History/background:

1. Originally, used curl with raw.github.com to get the latest version
of install-racket.sh.

2. But GitHub discourages using raw.github.com. So switched to using the
GitHub "releases" feature to make install-racket.sh available. But
that's a PITA: (a) Need to make each release, and furthermore (b)
there's no URL meaning "the latest version of the file to be
downloaded" for the .travis.yml scripts to use, so they tediously need
to be updated to fetch using the new URL for the new version.

3. OK, so now -- how about simply cloning the GitHub repo locally, and
using the travis-racket.sh from that (which of course will be the latest
version).