Overhaul on install-racket.sh.
Normalize urls as much as possible and to get minimal installs working
properly. This can probably be merged to master and drop the minimal
branch entirely.

I think there is still some cleanup that can happen. The various linux
flavors can probably be unified, if not parameterized.

This patch removes SCOPE_SNAPSHOT, as it seems to point to incredibly
stale content. RELEASE does too, but there is a note that it is only
valid during release phases.