Distinguish our own .travis.yml and example/.travis.yml
As a result, our own .travis.yml can focus on testing
install-racket.sh as of the current commit.

(I am still using test.sh and `for` loops. It's not immediately
obvious to me how to express that instead using a Travis job matrix --
or that doing so would be any clearer.)